feat: add right sidebar widget for Goodreads sync
Add a sidebar widget that shows feed URL status, last sync time, and a
manual "Sync Now" button. Extracts sync logic into shared src/sync.ts
module used by both the widget and the command palette action. Includes
Goodreads "g" logo (from UXWing, free for commercial use) as the widget
tab icon.
