+++ src/widgets/index.tsx
@@ -1,188 +1,23 @@
 import {
   declareIndexPlugin,
   type ReactRNPlugin,
-  PluginRem,
+  WidgetLocation,
 } from '@remnote/plugin-sdk';
 import '../style.css';
-import '../index.css'; // import <widget-name>.css
+import '../index.css';
 import { doError, doLog } from '../logging';
-import { GoodreadsBook, parseBooks } from '../parseRss';
-import { fetchRss } from '../fetchRss';
+import { performSync } from '../sync';
 
-const PARENT_REM_NAME = 'Goodreads Import';
-const BOOK_TAG_NAME = 'Book';
-const AUTHOR_TAG_NAME = 'Author';
-const AUTHORS_PROPERTY_NAME = 'Author(s)';
 const DEFAULT_SYNC_INTERVAL_MINUTES = 30;
 
 let syncIntervalId: ReturnType<typeof setInterval> | null = null;
 
-async function getOrCreateParentRem(plugin: ReactRNPlugin): Promise<PluginRem> {
-    const existingRem = await plugin.rem.findByName([PARENT_REM_NAME], null);
-    if (existingRem) {
-      doLog(`Parent Rem "${PARENT_REM_NAME}" found (${existingRem._id})`);
-      return existingRem;
-    }
-
-    const parentRem = await plugin.rem.createRem();
-    if (!parentRem) {
-      throw new Error(`Failed to create parent Rem "${PARENT_REM_NAME}"`);
-    }
-    await parentRem.setText([PARENT_REM_NAME]);
-    await parentRem.setIsDocument(true);
-    doLog(`Parent Rem "${PARENT_REM_NAME}" created (${parentRem._id})`);
-    return parentRem;
-}
-
-async function getOrCreateTagRem(
-  plugin: ReactRNPlugin,
-  tagName: string,
-  parentId: string
-): Promise<PluginRem> {
-    const existingRem = await plugin.rem.findByName([tagName], parentId);
-    if (existingRem) {
-      doLog(`Tag Rem "${tagName}" found (${existingRem._id})`);
-      return existingRem;
-    }
-
-    const tagRem = await plugin.rem.createRem();
-    if (!tagRem) {
-      throw new Error(`Failed to create tag Rem "${tagName}"`);
-    }
-    await tagRem.setText([tagName]);
-    await tagRem.setParent(parentId);
-    doLog(`Tag Rem "${tagName}" created (${tagRem._id})`);
-    return tagRem;
-}
-
-async function getOrCreateAuthorPropertyRem(
-  plugin: ReactRNPlugin,
-  bookTag: PluginRem
-): Promise<PluginRem> {
-    const existingRem = await plugin.rem.findByName([AUTHORS_PROPERTY_NAME], bookTag._id);
-    if (existingRem) {
-      doLog(`Author(s) property found (${existingRem._id})`);
-      return existingRem;
-    }
-
-    const propertyRem = await plugin.rem.createRem();
-    if (!propertyRem) {
-      throw new Error(`Failed to create Author(s) property Rem`);
-    }
-    await propertyRem.setText([AUTHORS_PROPERTY_NAME]);
-    await propertyRem.setParent(bookTag._id);
-    await propertyRem.setIsProperty(true);
-    doLog(`Author(s) property created (${propertyRem._id})`);
-    return propertyRem;
-}
-
-async function getOrCreateAuthorRem(
-  plugin: ReactRNPlugin,
-  authorName: string,
-  parentId: string,
-  authorTag: PluginRem
-): Promise<PluginRem> {
-    const existingRem = await plugin.rem.findByName([authorName], parentId);
-    if (existingRem) {
-      doLog(`Author Rem "${authorName}" found (${existingRem._id})`);
-      return existingRem;
-    }
-
-    const authorRem = await plugin.rem.createRem();
-    if (!authorRem) {
-      throw new Error(`Failed to create author Rem "${authorName}"`);
-    }
-    await authorRem.setText([authorName]);
-    await authorRem.setIsDocument(true);
-    await authorRem.setParent(parentId);
-    await authorRem.addTag(authorTag);
-    doLog(`Author Rem "${authorName}" created and tagged (${authorRem._id})`);
-    return authorRem;
-}
-
-interface CreateBookOptions {
-  plugin: ReactRNPlugin;
-  parentId: string;
-  bookTag: PluginRem;
-  authorTag: PluginRem;
-  authorsPropertyId: string;
-}
-
-async function createRemForBook(
-  book: GoodreadsBook,
-  { plugin, parentId, bookTag, authorTag, authorsPropertyId }: CreateBookOptions
-): Promise<PluginRem|undefined> {
-    const { title, author } = book;
-
-    // Check if a Rem with this title already exists under the parent
-    const existingRem = await plugin.rem.findByName([title], parentId);
-    if (existingRem) {
-      doLog(`Rem for "${title}" exists (${existingRem._id}), skipping`);
-      return;
-    } else {
-      doLog(`No rem for "${title}" found, creating`);
-    }
-
-    // Create new Rem for the book
-    const bookRem = await plugin.rem.createRem();
-    if (!bookRem) {
-      doError(`Failed to create Rem "${title}"`);
-      return;
-    }
-    doLog(`Rem created for "${title}" (${bookRem._id})`);
-    await bookRem.setText([title]);
-    await bookRem.setIsDocument(true);
-    await bookRem.setParent(parentId);
-
-    // Tag the book with the "Book" tag
-    await bookRem.addTag(bookTag);
-    doLog(`Rem tagged as Book for "${title}" (${bookRem._id})`);
-
-    // Create/find author and set the Author(s) property
-    if (author) {
-      const authorRem = await getOrCreateAuthorRem(plugin, author, parentId, authorTag);
-      const authorRichText = await plugin.richText.rem(authorRem).value();
-      await bookRem.setTagPropertyValue(authorsPropertyId, authorRichText);
-      doLog(`Author "${author}" linked to "${title}"`);
-    }
-
-    doLog(`Rem populated for "${title}" (${bookRem._id})`);
-    return bookRem;
-}
-
 async function fetchGoodreads(plugin: ReactRNPlugin) {
-  let remsCreated = 0;
   try {
-    // TODO: validate that the feedURL is a goodreads URL
-    const feedUrl: string = await plugin.settings.getSetting('feedUrl');
-    const xmlDoc = await fetchRss(feedUrl);
-
-    const cleanupTitle: boolean = await plugin.settings.getSetting('cleanupTitles');
-    const books = parseBooks(xmlDoc, {cleanupTitle});
-    doLog(`Found ${books.length} book(s) in feed`);
-
-    const parentRem = await getOrCreateParentRem(plugin);
-
-    // Create/find tag Rems
-    const bookTag = await getOrCreateTagRem(plugin, BOOK_TAG_NAME, parentRem._id);
-    const authorTag = await getOrCreateTagRem(plugin, AUTHOR_TAG_NAME, parentRem._id);
-
-    // Create/find the Author(s) property on the Book tag
-    const authorsProperty = await getOrCreateAuthorPropertyRem(plugin, bookTag);
-
-    // Process each book
-    for (const book of books) {
-      const rem = await createRemForBook(book, {
-        plugin,
-        parentId: parentRem._id,
-        bookTag,
-        authorTag,
-        authorsPropertyId: authorsProperty._id,
-      });
-      if(rem) remsCreated++;
-    }
-
-    await plugin.app.toast(`Goodreads sync complete. Found ${remsCreated} new book(s) (${books.length - remsCreated} existing)`);
+    const result = await performSync(plugin);
+    await plugin.app.toast(
+      `Goodreads sync complete. Found ${result.imported} new book(s) (${result.existing} existing)`
+    );
   } catch (error) {
     doError(`Error fetching Goodreads shelf: ${error}`);
     await plugin.app.toast('Error syncing Goodreads, check the console for additional details.');
@@ -233,7 +68,6 @@ async function onActivate(plugin: ReactRNPlugin) {
     defaultValue: DEFAULT_SYNC_INTERVAL_MINUTES,
   });
 
-  // Register a command to fetch books from Goodreads
   await plugin.app.registerCommand({
     id: 'fetch-goodreads-shelf',
     name: 'Fetch Books from Goodreads Shelf',
@@ -242,7 +76,12 @@ async function onActivate(plugin: ReactRNPlugin) {
     }
   });
 
-  // Start periodic sync based on setting
+  await plugin.app.registerWidget('goodreads_widget', WidgetLocation.RightSidebar, {
+    dimensions: { height: 'auto', width: '100%' },
+    widgetTabTitle: 'Goodreads',
+    widgetTabIcon: `${plugin.rootURL}logo.svg`,
+  });
+
   const syncInterval: number = await plugin.settings.getSetting('syncIntervalMinutes');
   startPeriodicSync(plugin, syncInterval);
 }
